test: check if git is present before running the tests

+++ test/parallel/test-fs-rm.js
@@ -16,6 +16,10 @@ let count = 0;
 const nextDirPath = (name = 'rm') =>
   path.join(tmpdir.path, `${name}-${count++}`);
 
+const isGitPresent = (() => {
+  try { execSync('git --version'); return true; } catch { return false; }
+})();
+
 function makeNonEmptyDirectory(depth, files, folders, dirname, createSymLinks) {
   fs.mkdirSync(dirname, { recursive: true });
   fs.writeFileSync(path.join(dirname, 'text.txt'), 'hello', 'utf8');
@@ -132,7 +136,7 @@ function removeAsync(dir) {
 
 // Removing a .git directory should not throw an EPERM.
 // Refs: https://github.com/isaacs/rimraf/issues/21.
-{
+if (isGitPresent) {
   const gitDirectory = nextDirPath();
   fs.mkdirSync(gitDirectory);
   execSync(`git -C ${gitDirectory} init`);
@@ -192,7 +196,7 @@ function removeAsync(dir) {
 
 // Removing a .git directory should not throw an EPERM.
 // Refs: https://github.com/isaacs/rimraf/issues/21.
-{
+if (isGitPresent) {
   const gitDirectory = nextDirPath();
   fs.mkdirSync(gitDirectory);
   execSync(`git -C ${gitDirectory} init`);
@@ -253,13 +257,15 @@ function removeAsync(dir) {
 
 // Removing a .git directory should not throw an EPERM.
 // Refs: https://github.com/isaacs/rimraf/issues/21.
-(async () => {
-  const gitDirectory = nextDirPath();
-  fs.mkdirSync(gitDirectory);
-  execSync(`git -C ${gitDirectory} init`);
-  await fs.promises.rm(gitDirectory, { recursive: true });
-  assert.strictEqual(fs.existsSync(gitDirectory), false);
-})().then(common.mustCall());
+if (isGitPresent) {
+  (async () => {
+    const gitDirectory = nextDirPath();
+    fs.mkdirSync(gitDirectory);
+    execSync(`git -C ${gitDirectory} init`);
+    await fs.promises.rm(gitDirectory, { recursive: true });
+    assert.strictEqual(fs.existsSync(gitDirectory), false);
+  })().then(common.mustCall());
+}
 
 // Test input validation.
 {
